Take the shop totals in long on both sides of the counter
RequestBuyItem bounds count * price per item, but on the price before tax, and
the price is multiplied by (1 + castleTax + baseTax) on the very next line. The
running total was an int, so the taxed total could wrap, and a wrapped total
passes the MAX_ADENA test it is there to fail. What singles it out is the weight
test twenty lines down, which already reads "weight > Integer.MAX_VALUE" -- a
condition that can only ever be true if the product feeding it is widened, and it
was not.

RequestSellItem adds to the total before testing it, and both the product and the
total were int. Its per item test catches one item worth more than the limit, but
not the sum of several. The test now runs first, which also keeps the goods from
moving for the item that breaks the limit.

Measured rather than assumed. MaxAdena ships at 2,000,000,000 and the dearest
product in the 6357 priced buylist entries is 10,000,000, so the per item test
admits counts whose taxed total does not fit in an int at the 10 to 15 percent
baseTax the shipped merchant data carries.

Dismissed in the same pass: RequestPreviewItem accumulates WearPrice, which ships
at 10, over at most 100 items; RateSiegeGuardsPrice ships at 1, so the price
scaling above is a no-op; and both sell paths reject a per item count below one
when they read the packet, so neither division can be by zero.

+++ java/org/l2jmobius/gameserver/network/clientpackets/RequestBuyItem.java
@@ -159,7 +159,12 @@ public class RequestBuyItem extends ClientPacket
 			}
 		}
 		
-		int subTotal = 0;
+		// The per item check below bounds count * price, but it runs on the price before
+		// tax and the price grows right after it. Accumulating in int let the taxed total
+		// wrap, and the wrapped value passes the MAX_ADENA test it should fail. The weight
+		// test twenty lines down already reads "weight > Integer.MAX_VALUE", which can only
+		// ever be true if the product feeding it is widened first.
+		long subTotal = 0;
 		
 		// Check for buylist validity and calculates summary values
 		long slots = 0;
@@ -217,14 +222,14 @@ public class RequestBuyItem extends ClientPacket
 			
 			// first calculate price per item with tax, then multiply by count
 			price = (int) (price * (1 + castleTaxRate + baseTaxRate));
-			subTotal += i.getCount() * price;
+			subTotal += (long) i.getCount() * price;
 			if (subTotal > MAX_ADENA)
 			{
 				PunishmentManager.handleIllegalPlayerAction(player, "Warning!! Character " + player.getName() + " of account " + player.getAccountName() + " tried to purchase over " + MAX_ADENA + " adena worth of goods.", GeneralConfig.DEFAULT_PUNISH);
 				return;
 			}
 			
-			weight += i.getCount() * product.getItem().getWeight();
+			weight += (long) i.getCount() * product.getItem().getWeight();
 			if (player.getInventory().getItemByItemId(product.getItemId()) == null)
 			{
 				slots++;
@@ -246,7 +251,7 @@ public class RequestBuyItem extends ClientPacket
 		}
 		
 		// Charge buyer and add tax to castle treasury if not owned by npc clan
-		if ((subTotal < 0) || !player.reduceAdena(ItemProcessType.BUY, subTotal, player.getLastFolkNPC(), false))
+		if ((subTotal < 0) || !player.reduceAdena(ItemProcessType.BUY, (int) subTotal, player.getLastFolkNPC(), false))
 		{
 			player.sendPacket(SystemMessageId.YOU_DO_NOT_HAVE_ENOUGH_ADENA);
 			player.sendPacket(ActionFailed.STATIC_PACKET);
